refactor: re-organize file structures
