+++ src/github/embeds/PushEmbed.ts
@@ -1,7 +1,7 @@
-import { ApplyOptions } from "@snowcrystals/iglo";
-import { GitHubEmbed, type GitHubEmbedOptions } from "../structures/GitHubEmbed.js";
+import { GitHubEmbed, type GitHubEmbedOptions } from "../lib/embed/structures/GitHubEmbed.js";
 import type { PushEvent } from "@octokit/webhooks-types";
 import type { EmbedBuilder } from "discord.js";
+import { ApplyOptions } from "../lib/embed/decorators.js";
 
 @ApplyOptions<GitHubEmbedOptions>({ name: "push" })
 export default class extends GitHubEmbed {
