+++ src/github/lib/GitHubManager.ts
@@ -1,5 +1,5 @@
 import type GitCordClient from "../../lib/GitCordClient.js";
-import GitHubEmbedLoader from "./GitHubEmbedLoader.js";
+import GitHubEmbedLoader from "./embed/GitHubEmbedLoader.js";
 import GitHubWebhookManager from "./webhook/GitHubWebhookManager.js";
 
 export default class GitHubManager {
